+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [0.2.0] - 2021-07-07
 ### Changed
 - Move Colors class to colors.py.
+- Move load and save json functions to files.py.
 ### Fix
 - Incorrect type of tags in tasks.py.